test: move fnc to other file, console warn instead of error
