+++ ui/src/components/CheckBoxComponent/CheckBoxComponent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Switch from '@splunk/react-ui/Switch';
-import { isFalse } from '../../util/util';
+import { isFalse } from '../../util/considerFalseAndTruthy';
 
 interface CheckBoxComponentProps {
     value: 0 | 1 | boolean;
