+++ ui/src/util/util.ts
@@ -2,10 +2,6 @@ import { TOAST_TYPES } from '@splunk/react-toast-notifications/ToastConstants';
 import Toaster, { makeCreateToast } from '@splunk/react-toast-notifications/Toaster';
 import { GlobalConfig, GlobalConfigSchema } from '../types/globalConfig/globalConfig';
 import { invariant } from './invariant';
-import {
-    AcceptableFormValue,
-    AcceptableFormValueOrNullish,
-} from '../types/components/shareableTypes';
 
 interface AppData {
     app: string;
@@ -26,22 +22,6 @@ export function getMetaInfo() {
     return {
         appData,
     };
-}
-
-export function isFalse(value: AcceptableFormValueOrNullish) {
-    return (
-        value === null ||
-        value === undefined ||
-        ['0', 'FALSE', 'F', 'N', 'NO', 'NONE', ''].includes(value.toString().toUpperCase())
-    );
-}
-
-export function isTrue(value: AcceptableFormValueOrNullish) {
-    return (
-        value !== null &&
-        value !== undefined &&
-        ['1', 'TRUE', 'T', 'Y', 'YES'].includes(value.toString().toUpperCase())
-    );
 }
 
 export function generateEndPointUrl(name: string) {
@@ -145,12 +125,3 @@ export function filterResponse(
     }
     return newItems;
 }
-
-/**
- * since splunk does not oeprate on booleans and numbers, but onyl strings
- * here we use mapping to enable compering and oeprating on them
- * @param value value used for mapping
- * @returns maps truthy values into 1 and false into 0, does not midify rest
- */
-export const getValueConsiderTruthyFalse = (value: AcceptableFormValue) =>
-    (isFalse(value) && '0') || (isTrue(value) && '1') || value;
